% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl.R
\name{addGraph}
\alias{addGraph}
\title{Mutates geom_smooth lm plots onto tbl--needs $varData}
\usage{
addGraph(...statEvalNonTbl, ....formula)
}
\arguments{
\item{....formula}{}
}
\value{
mutated tbl
}
\description{
Mutates geom_smooth lm plots onto tbl--needs $varData
}
