% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl1.R
\name{addGraph1}
\alias{addGraph1}
\title{Mutates geom_smooth lm plots, for already-summarized data, onto tbl--needs $varData}
\usage{
addGraph1(...statEvalNonTbl, ....formula)
}
\arguments{
\item{....formula}{}
}
\value{
mutated tbl
}
\description{
Mutates geom_smooth lm plots, for already-summarized data, onto tbl--needs $varData
}
