% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl12.R
\name{addGraph12}
\alias{addGraph12}
\title{Mutates quadratic plots, for already-summarized data, onto tbl--needs $varData}
\usage{
addGraph12(...statEvalNonTbl2, ....formula)
}
\arguments{
\item{....formula}{}
}
\value{
mutated tbl
}
\description{
Mutates quadratic plots, for already-summarized data, onto tbl--needs $varData
}
