% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl2.R
\name{addGraph2}
\alias{addGraph2}
\title{Mutates quadratic plots onto tbl--needs $varData}
\usage{
addGraph2(...statEvalNonTbl2, ....formula)
}
\arguments{
\item{....formula}{}
}
\value{
mutated tbl
}
\description{
Mutates quadratic plots onto tbl--needs $varData
}
