% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl.R
\name{addStatEval}
\alias{addStatEval}
\title{Mutates lm assumptions eval onto tbl--needs $varData, $statTest, $statPrint}
\usage{
addStatEval(...statFitTbl, ....formula)
}
\arguments{
\item{....formula}{}
}
\value{
mutated tbl
}
\description{
Mutates lm assumptions eval onto tbl--needs $varData, $statTest, $statPrint
}
