% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl2.R
\name{addStatEval2}
\alias{addStatEval2}
\title{Mutates key lm info onto tbl--needs $statPrint}
\usage{
addStatEval2(...statFitTbl2, ....formula2)
}
\arguments{
\item{....formula2}{}
}
\value{

}
\description{
Mutates key lm info onto tbl--needs $statPrint
}
