% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonStatsTbl.R
\name{addStatEvalNon}
\alias{addStatEvalNon}
\title{Mutates key glm values onto tbl--needs $isModelOK, $statTestPois, $statTestGamma}
\usage{
addStatEvalNon(...statFitNonTbl)
}
\arguments{
\item{...statFitNonTbl}{}
}
\value{
mutated tbl
}
\description{
Mutates key glm values onto tbl--needs $isModelOK, $statTestPois, $statTestGamma
}
