% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl.R
\name{addStatFitNon}
\alias{addStatFitNon}
\title{Mutates glm summary output onto tbl--needs $varData}
\usage{
addStatFitNon(...statEvalTbl, ....formula)
}
\arguments{
\item{....formula}{}
}
\value{
mutated tbl
}
\description{
Mutates glm summary output onto tbl--needs $varData
}
