% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonStatsTbl.R
\name{addStatFitNonNP}
\alias{addStatFitNonNP}
\title{Mutates kruskal_test summary output onto tbl--needs $varData, $isModelOK}
\usage{
addStatFitNonNP(...statEvalTbl, ....formula)
}
\arguments{
\item{....formula}{}
}
\value{
mutated tbl
}
\description{
Mutates kruskal_test summary output onto tbl--needs $varData, $isModelOK
}
