% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{dotGraph}
\alias{dotGraph}
\title{Sets cleaner ggplot style, scatter default}
\usage{
dotGraph(
  ..varData,
  ..var,
  ..x,
  ..y,
  ..xlab,
  ..ylab,
  ..addBins = F,
  ..cleanData = NULL,
  ..addLines = F,
  ..log10Axes = F,
  ..useGroups = NULL,
  ..groupTitle = NULL,
  ..addCenters = F,
  ..dark = F
)
}
\arguments{
\item{..useGroups}{}
}
\value{
scatter ggplot
}
\description{
Sets cleaner ggplot style, scatter default
}
