% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordTbl.R
\name{getOrdVarTbl}
\alias{getOrdVarTbl}
\title{Organize ordination attributes into plot-able tibble}
\usage{
getOrdVarTbl(...commTbl, ...metaTbl)
}
\arguments{
\item{...metaTbl}{}
}
\value{
tibble of coordinates and sample attributes
}
\description{
Organize ordination attributes into plot-able tibble
}
