% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl.R
\name{getStatsTbl}
\alias{getStatsTbl}
\title{Gather results of basic regression, assumption checks, & generalized versions}
\usage{
getStatsTbl(...nestedVarDataTbl, ...formula)
}
\arguments{
\item{...formula}{}
}
\value{
large tibble
}
\description{
Gather results of basic regression, assumption checks, & generalized versions
}
