% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl1.R
\name{getStatsTbl1}
\alias{getStatsTbl1}
\title{Gather results, for already-summarized data, of basic regression, assumptions checks and generalized options}
\usage{
getStatsTbl1(...nestedVarDataTbl, ...formula)
}
\arguments{
\item{...formula}{}
}
\value{
large tibble
}
\description{
Gather results, for already-summarized data, of basic regression, assumptions checks and generalized options
}
