% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl12.R
\name{getStatsTbl12}
\alias{getStatsTbl12}
\title{Gather results, for already-summarized data, of quadratic regression w/o assumption or generalized check/options}
\usage{
getStatsTbl12(...nestedVarDataTbl, ...formula)
}
\arguments{
\item{...formula}{}
}
\value{
large tibble
}
\description{
Gather results, for already-summarized data, of quadratic regression w/o assumption or generalized check/options
}
