% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl2.R
\name{getStatsTbl2}
\alias{getStatsTbl2}
\title{Gather results of quadratic regression w/o assumption or generalized check/options}
\usage{
getStatsTbl2(...nestedVarDataTbl, ...formula2)
}
\arguments{
\item{...formula2}{}
}
\value{
large tibble
}
\description{
Gather results of quadratic regression w/o assumption or generalized check/options
}
