% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{runFits}
\alias{runFits}
\title{Runs distribution fits through sample list}
\usage{
runFits(...nestedTbl)
}
\arguments{
\item{...nestedTbl}{}
}
\value{
tbl
}
\description{
Runs distribution fits through sample list
}
