% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl.R
\name{statFitTbl}
\alias{statFitTbl}
\title{Mutates lm and summary output onto tbl--needs $varData, $statTest}
\usage{
statFitTbl(...nestedVarDataTbl, ....formula)
}
\arguments{
\item{....formula}{}
}
\value{
mutated tbl
}
\description{
Mutates lm and summary output onto tbl--needs $varData, $statTest
}
