% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTbl1.R
\name{statFitTbl1}
\alias{statFitTbl1}
\title{Mutates lm and summary output, of already-summarized data, onto tbl--needs $varData1, $statTest}
\usage{
statFitTbl1(...nestedVarDataTbl, ....formula)
}
\arguments{
\item{....formula}{}
}
\value{
mutated tbl
}
\description{
Mutates lm and summary output, of already-summarized data, onto tbl--needs $varData1, $statTest
}
