% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poweRlawTest.R
\name{CheckPoweRlaw}
\alias{CheckPoweRlaw}
\title{Gather results of non-linear distribution fits to validate a power law fit}
\usage{
CheckPoweRlaw(RealFreq, xMin = NULL, xMax = 1e+05, nSims = 100)
}
\arguments{
\item{RealFreq}{}
}
\value{
"results" data frame
}
\description{
Gather results of non-linear distribution fits to validate a power law fit
}
