% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{dotGraph}
\alias{dotGraph}
\title{Sets cleaner ggplot style, scatter default}
\usage{
dotGraph(
  ..varData,
  ..var,
  ..x,
  ..y,
  ..xlab,
  ..ylab,
  ..addGroups = F,
  ..cleanData = NULL
)
}
\arguments{
\item{..cleanData}{}
}
\value{

}
\description{
Sets cleaner ggplot style, scatter default
}
